Create Selection sort
Selection sort is a sorting algorithm that sorts an array by repeatedly finding the minimum element and putting them in ascending order. Selection sort in python does not require any extra space, so the sorting is done in place. The average time complexity of selection sort in python is  O(N^2)