Create Heap sort
Heap sort processes the elements by creating the min-heap or max-heap using the elements of the given array. Min-heap or max-heap represents the ordering of array in which the root element represents the minimum or maximum element of the array.

Heap sort basically recursively performs two main operations -

Build a heap H, using the elements of array.
Repeatedly delete the root element of the heap formed in 1st phase.